Make sure to detect 1900 and 1904 date system correctly
A spreadsheet uses the 1904 date system if the `date1904` attribute
of `workbook.xml` of the spreadsheet is set to either `true` or `1`.
The counterpart is `false` and `0`
Example:

```
<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
<workbook xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships"><fileVersion appName="xl" lastEdited="5" lowestEdited="5" rupBuild="9303"/><workbookPr date1904="true" defaultThemeVersion="124226"/>
```

The previous detection system was checking only for `false` and not for `0`.
This commits changes that to make sure both values are checked.
The condition is also reversed for better readability.

See https://docs.microsoft.com/en-us/office/troubleshoot/excel/1900-and-1904-date-system